requirements.txt
streamlit==1.39.0
pandas==2.2.2
langchain-openai==0.1.23
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv
tavily-python
aiohttp
rapidfuzz